+++ aspirant_application/Graphics.FloorInventory.cpp
@@ -85,9 +85,9 @@ namespace graphics::FloorInventory
 			auto color = (index == inventoryIndex) ? (activeColor) : (inactiveColor);
 			if (dropShadow)
 			{
-				graphics::Fonts::Read(font).value().WriteText(renderer, { x + dropShadowX,y + dropShadowY }, ss.str(), dropShadowColor, graphics::HorizontalAlignment::LEFT);
+				graphics::Fonts::WriteText(font, renderer, { x + dropShadowX,y + dropShadowY }, ss.str(), dropShadowColor, graphics::HorizontalAlignment::LEFT);
 			}
-			graphics::Fonts::Read(font).value().WriteText(renderer, { x,y }, ss.str(), color, graphics::HorizontalAlignment::LEFT);
+			graphics::Fonts::WriteText(font, renderer, { x,y }, ss.str(), color, graphics::HorizontalAlignment::LEFT);
 			y += rowHeight;
 			index++;
 		}
@@ -96,9 +96,9 @@ namespace graphics::FloorInventory
 		{
 			if (dropShadow)
 			{
-				graphics::Fonts::Read(font).value().WriteText(renderer, { x + dropShadowX,y + dropShadowY }, "(nothing)", dropShadowColor, graphics::HorizontalAlignment::LEFT);
+				graphics::Fonts::WriteText(font, renderer, { x + dropShadowX,y + dropShadowY }, "(nothing)", dropShadowColor, graphics::HorizontalAlignment::LEFT);
 			}
-			graphics::Fonts::Read(font).value().WriteText(renderer, { x,y }, "(nothing)", inactiveColor, graphics::HorizontalAlignment::LEFT);
+			graphics::Fonts::WriteText(font, renderer, { x,y }, "(nothing)", inactiveColor, graphics::HorizontalAlignment::LEFT);
 		}
 	}
 }
