+++ aspirant_application/Graphics.Fonts.cpp
@@ -1,5 +1,110 @@
 #include "Graphics.Fonts.h"
 #include "Data.JSON.h"
+#include "Graphics.Sprites.h"
+#include <sstream>
+#include "Graphics.Colors.h"
+namespace graphics
+{
+	class Font
+	{
+	private:
+		const nlohmann::json& model;
+		std::optional<graphics::Sprite> GetGlyphSprite(char) const;
+		void WriteTextCentered(std::shared_ptr<SDL_Renderer>, const common::XY<int>&, const std::string&, const std::string&) const;
+		common::XY<int> WriteTextLeft(std::shared_ptr<SDL_Renderer>, const common::XY<int>&, const std::string&, const std::string&) const;
+		void WriteTextRight(std::shared_ptr<SDL_Renderer>, const common::XY<int>&, const std::string&, const std::string&) const;
+		common::XY<int> WriteGlyph(std::shared_ptr<SDL_Renderer>, const common::XY<int>&, char, const std::string&) const;
+	public:
+		Font(const nlohmann::json&);
+		void WriteText(std::shared_ptr<SDL_Renderer>, const common::XY<int>&, const std::string&, const std::string&, const HorizontalAlignment&) const;
+	};
+
+	Font::Font
+	(
+		const nlohmann::json& model
+	)
+		: model(model)
+	{
+	}
+
+	std::optional<graphics::Sprite> Font::GetGlyphSprite(char ch) const
+	{
+		std::stringstream ss;
+		ss << (int)ch;
+		if (model.count(ss.str()) > 0)
+		{
+			return Sprites::Read(model[ss.str()]);
+		}
+		else
+		{
+			return std::nullopt;
+		}
+	}
+
+	common::XY<int> Font::WriteGlyph(std::shared_ptr<SDL_Renderer> renderer, const common::XY<int>& xy, char ch, const std::string& color) const
+	{
+		const auto& sprite = GetGlyphSprite(ch);
+		sprite.value().Draw(renderer, xy, ::graphics::Colors::Read(color));
+		return common::XY(xy.GetX() + sprite.value().GetWidth(), xy.GetY());
+	}
+
+	common::XY<int> Font::WriteTextLeft(std::shared_ptr<SDL_Renderer> renderer, const common::XY<int>& xy, const std::string& text, const std::string& color) const
+	{
+		common::XY<int> temp = xy;
+		for (auto ch : text)
+		{
+			temp = WriteGlyph(renderer, temp, ch, color);
+		}
+		return temp;
+	}
+
+	void Font::WriteTextCentered(std::shared_ptr<SDL_Renderer> renderer, const common::XY<int>& xy, const std::string& text, const std::string& color) const
+	{
+		int width = 0;
+		for (auto ch : text)
+		{
+			const auto& sprite = GetGlyphSprite(ch);
+			width += sprite.value().GetWidth();
+		}
+		auto adjustedXY = common::XY<int>(xy.GetX() - width / 2, xy.GetY());
+		WriteTextLeft(renderer, adjustedXY, text, color);
+	}
+
+	void Font::WriteTextRight(std::shared_ptr<SDL_Renderer> renderer, const common::XY<int>& xy, const std::string& text, const std::string& color) const
+	{
+		int width = 0;
+		for (auto ch : text)
+		{
+			const auto& sprite = GetGlyphSprite(ch);
+			width += sprite.value().GetWidth();
+		}
+		auto adjustedXY = common::XY<int>(xy.GetX() - width, xy.GetY());
+		WriteTextLeft(renderer, adjustedXY, text, color);
+	}
+
+	void Font::WriteText
+	(
+		std::shared_ptr<SDL_Renderer> renderer,
+		const common::XY<int>& xy,
+		const std::string& text,
+		const std::string& color,
+		const HorizontalAlignment& alignment
+	) const
+	{
+		switch (alignment)
+		{
+		case HorizontalAlignment::LEFT:
+			WriteTextLeft(renderer, xy, text, color);
+			break;
+		case HorizontalAlignment::RIGHT:
+			WriteTextRight(renderer, xy, text, color);
+			break;
+		case HorizontalAlignment::CENTER:
+			WriteTextCentered(renderer, xy, text, color);
+			break;
+		}
+	}
+}
 namespace graphics::Fonts
 {
 	static std::map<std::string, nlohmann::json> table;
@@ -9,7 +114,7 @@ namespace graphics::Fonts
 		table[identifier] = data::JSON::Load(properties);
 	}
 
-	std::optional<graphics::Font> Read(const std::string& key)
+	static std::optional<graphics::Font> Read(const std::string& key)
 	{
 		if (table.count(key) > 0)
 		{
@@ -30,4 +135,12 @@ namespace graphics::Fonts
 		}
 	}
 
+	void WriteText(const std::string& fontname, std::shared_ptr<SDL_Renderer> renderer, const common::XY<int>& xy, const std::string& text, const std::string& color, const HorizontalAlignment& alignment)
+	{
+		auto font = Read(fontname);
+		if (font)
+		{
+			font.value().WriteText(renderer, xy, text, color, alignment);
+		}
+	}
 }
