+++ aspirant_application/Graphics.Menu.cpp
@@ -24,7 +24,8 @@ namespace graphics::Menu
 			int y = menuItem[common::Properties::Y];
 			if (dropShadow)
 			{
-				Fonts::Read(font).value().WriteText(
+				Fonts::WriteText(
+					font,
 					renderer, 
 					common::XY<int>(x + dropShadowX, y + dropShadowY), 
 					text, 
@@ -32,7 +33,7 @@ namespace graphics::Menu
 					horizontalAlignment);
 			}
 			std::string color = (currentIndex == index) ? (activeColor) : (inactiveColor);
-			Fonts::Read(font).value().WriteText(renderer, common::XY<int>(x, y), text, color, horizontalAlignment);
+			Fonts::WriteText(font, renderer, common::XY<int>(x, y), text, color, horizontalAlignment);
 			currentIndex++;
 		}
 	}
