+++ aspirant_application/aspirant_application.vcxproj
@@ -216,7 +216,6 @@
     <ClCompile Include="Common.Audio.cpp" />
     <ClCompile Include="State.Splash.cpp" />
     <ClCompile Include="Graphics.Sprite.cpp" />
-    <ClCompile Include="Graphics.Font.cpp" />
     <ClCompile Include="Graphics.Sprites.cpp" />
     <ClCompile Include="State.Start.cpp" />
     <ClCompile Include="Graphics.Image.cpp" />
@@ -280,7 +279,6 @@
     <ClInclude Include="Common.RNG.h" />
     <ClInclude Include="Common.Audio.h" />
     <ClInclude Include="Graphics.Sprite.h" />
-    <ClInclude Include="Graphics.Font.h" />
     <ClInclude Include="Graphics.Sprites.h" />
     <ClInclude Include="Graphics.Textures.h" />
     <ClInclude Include="UIState.h" />
